@@ -11,17 +11,17 @@
     "start": "next start"
   },
   "engines": {
-    "node": ">=6.10.0"
+    "node": ">=8.0.0"
   },
   "dependencies": {
+    "@material-ui/core": "^3.1.1",
+    "@material-ui/icons": "3.0.1",
     "date-fns": "^1.28.3",
-    "debug": "^3.1.0",
-    "deepstream.io-client-js": "^2.1.4",
+    "debug": "^4.0.1",
+    "deepstream.io-client-js": "latest",
     "jss": "^9.8.1",
-    "material-ui": "^1.0.0-beta.38",
-    "material-ui-icons": "^1.0.0-beta.17",
-    "next": "^5.1.0",
-    "next-redux-wrapper": "^1.3.5",
+    "next": "^7.0.0",
+    "next-redux-wrapper": "^2.0.0",
     "prop-types": "^15.5.10",
     "ramda": "^0.25.0",
     "react": "^16.3.0",
@@ -30,21 +30,24 @@
     "react-jss": "^8.3.5",
     "react-motion": "^0.5.0",
     "react-redux": "^5.0.4",
-    "recompose": "^0.26.0",
-    "redux": "^3.7.2",
+    "recompose": "^0.30.0",
+    "redux": "^4.0.0",
     "redux-devtools-extension": "^2.13.2",
     "redux-logger": "^3.0.1",
-    "redux-observable": "0.18.0",
-    "rxjs": "^5.5.7",
+    "redux-observable": "1.0.0",
+    "rxjs": "^6.3.3",
     "shortid": "^2.2.8",
     "supports-color": "^5.0.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
+    "@babel/cli": "7.1.0",
+    "@babel/core": "7.1.0",
+    "@babel/preset-env": "7.1.0",
+    "@babel/preset-flow": "^7.0.0",
     "babel-plugin-transform-define": "^1.2.0",
-    "babel-preset-env": "^1.6.0",
-    "eslint-plugin-flowtype": "^2.35.1",
-    "flow-bin": "^0.68.0",
-    "webpack-bundle-analyzer": "^2.9.0"
+    "eslint-config-react-tools": "^1.2.5",
+    "eslint-plugin-flowtype": "^2.47.1",
+    "flow-bin": "^0.81.0",
+    "webpack-bundle-analyzer": "^3.0.2"
   }
 }
