@@ -1,12 +1,11 @@
 // @flow
 
-import type { Action, Client } from '../../types';
+import type { Action, Client } from '../../types'
 
-import { Observable } from 'rxjs/Observable';
-import { map } from 'ramda';
+import { Observable } from 'rxjs'
 
 function getEntriesFromRecordList(client: Client, listname: string) {
-  const lst = client.record.getList(`smartHome/${listname}`);
+  const lst = client.record.getList(`smartHome/${listname}`)
 
   return Observable.bindCallback(cb => lst.whenReady(cb))().mergeMap(lst =>
     Observable.from(
@@ -17,7 +16,7 @@ function getEntriesFromRecordList(client: Client, listname: string) {
     )
       .concatAll()
       .toArray()
-  );
+  )
 }
 
-export { getEntriesFromRecordList };
+export { getEntriesFromRecordList }
