@@ -1,19 +1,20 @@
 // @flow
-import * as React from 'react';
-import Link from 'next/link';
-import { connect } from 'react-redux';
 
-import { withStyles } from 'material-ui/styles';
-import AppBar from 'material-ui/AppBar';
-import Toolbar from 'material-ui/Toolbar';
-import Button from 'material-ui/Button';
-import IconButton from 'material-ui/IconButton';
-import HomeIcon from 'material-ui-icons/Home';
-import MainIcon from 'material-ui-icons/Dashboard';
-import ConnIndicator from './ConnectionIndicator';
-import { grey, indigo } from 'material-ui/colors';
+import * as React from 'react'
+import Link from 'next/link'
+import { connect } from 'react-redux'
 
-import { compose } from 'ramda';
+import { withStyles } from '@material-ui/core/styles'
+import AppBar from '@material-ui/core/AppBar'
+import Toolbar from '@material-ui/core/Toolbar'
+import Button from '@material-ui/core/Button'
+import IconButton from '@material-ui/core/IconButton'
+import HomeIcon from '@material-ui/icons/Home'
+import MainIcon from '@material-ui/icons/Dashboard'
+import ConnIndicator from './ConnectionIndicator'
+import { grey, indigo } from '@material-ui/core/colors'
+
+import { compose } from 'ramda'
 
 type AppBarProps = {
   conn: {
@@ -21,7 +22,7 @@ type AppBarProps = {
     error: string,
   },
   classes: Object,
-};
+}
 
 const styles = {
   root: {
@@ -41,10 +42,14 @@ const styles = {
   flexFromHere: {
     flex: 1,
   },
-};
+}
 
 const MainAppBar = (props: AppBarProps) => {
-  const { classes, conn: { state: connState, error: connErr } } = props;
+  const {
+    classes,
+    conn: { state: connState, error: connErr },
+  } = props
+
   return (
     <div className={classes.root}>
       <AppBar className={classes.appBar}>
@@ -75,10 +80,10 @@ const MainAppBar = (props: AppBarProps) => {
         </Toolbar>
       </AppBar>
     </div>
-  );
-};
+  )
+}
 
 export default compose(
   withStyles(styles),
   connect(state => ({ conn: state.app.connection }))
-)(MainAppBar);
+)(MainAppBar)
