@@ -1,16 +1,16 @@
 // @flow
-import * as React from 'react';
-import { connect } from 'react-redux';
-import { toggleShowOnlyActive } from '../lib/app/actions';
-import Switch from 'material-ui/Switch';
+import * as React from 'react'
+import { connect } from 'react-redux'
+import { toggleShowOnlyActive } from '../lib/app/actions'
+import Switch from '@material-ui/core/Switch'
 
 type Props = {
   showActive: boolean,
   onToggleClick: Function,
-};
+}
 
 const showOnlyActiveToggle = (props: Props) => {
-  const { showActive, onToggleClick } = props;
+  const { showActive, onToggleClick } = props
   return (
     <div>
       <Switch
@@ -19,16 +19,16 @@ const showOnlyActiveToggle = (props: Props) => {
         onChange={(event, checked) => onToggleClick(!checked)}
       />
     </div>
-  );
-};
+  )
+}
 
 const mapDispatchToProps = dispatch => ({
   onToggleClick: (toggleChecked: boolean) => {
-    dispatch(toggleShowOnlyActive(toggleChecked));
+    dispatch(toggleShowOnlyActive(toggleChecked))
   },
-});
+})
 
 export default connect(
   state => ({ showActive: state.app.prefs.showOnlyActive }),
   mapDispatchToProps
-)(showOnlyActiveToggle);
+)(showOnlyActiveToggle)
