@@ -2,20 +2,17 @@
 
 // Container-component for all dashboard logic
 
-import type { Dispatch, KnxAddress, Prefs, Rooms, State } from '../../types';
+import type { Dispatch, KnxAddress, Prefs, Rooms, State } from '../../types'
 
-import * as React from 'react';
-import { connect } from 'react-redux';
-import { withStyles } from 'material-ui/styles';
-import Grid from 'material-ui/Grid';
-import OverviewLights from './OverviewLights';
+import * as React from 'react'
+import { connect } from 'react-redux'
+import { withStyles } from '@material-ui/core/styles'
+import Grid from '@material-ui/core/Grid'
+import OverviewLights from './OverviewLights'
 
-import {
-  toggleAddrVal,
-  onlyManuallySwitchedLights,
-} from '../../lib/shared/address-utils';
+import { toggleAddrVal, onlyManuallySwitchedLights } from '../../lib/shared/address-utils'
 
-import { compose } from 'ramda';
+import { compose } from 'ramda'
 
 type Props = {
   addresses: Array<KnxAddress>,
@@ -23,7 +20,7 @@ type Props = {
   rooms: Rooms,
   classes: Object,
   dispatch: Dispatch,
-};
+}
 
 // TODO
 const styles = theme => ({
@@ -33,14 +30,14 @@ const styles = theme => ({
   control: {
     padding: theme.spacing.unit * 2,
   },
-});
+})
 
 const Dashboard = ({ addresses, dispatch, classes }: Props) => {
   const onLightSwitch = addr =>
     dispatch({
       type: 'WRITE_GROUP_ADDRESS',
       addr: toggleAddrVal(addr),
-    });
+    })
 
   return (
     <Grid container className={classes.root}>
@@ -52,8 +49,8 @@ const Dashboard = ({ addresses, dispatch, classes }: Props) => {
         />
       </Grid>
     </Grid>
-  );
-};
+  )
+}
 
 export default compose(
   connect((state: State) => ({
@@ -61,4 +58,4 @@ export default compose(
     prefs: state.app.prefs,
   })),
   withStyles(styles)
-)(Dashboard);
+)(Dashboard)
