@@ -1,14 +1,18 @@
-import Avatar from 'material-ui/Avatar';
-import { green, grey } from 'material-ui/colors';
-import FermenterIcon from 'material-ui-icons/CallToAction';
-import { withStyles } from 'material-ui/styles';
+// @flow
 
-import { compose } from 'ramda';
+import * as React from 'react'
+import Avatar from '@material-ui/core/Avatar'
+import { green, grey } from '@material-ui/core/colors'
+import { withStyles } from '@material-ui/core/styles'
+
+import FermenterIcon from '@material-ui/icons/CallToAction'
+
+import { compose } from 'ramda'
 
 type Props = {
   classes: Object,
   isOn: boolean,
-};
+}
 
 const fermenterIconStyles = {
   fermenterIconOn: {
@@ -19,12 +23,12 @@ const fermenterIconStyles = {
     color: '#fff',
     background: grey[500],
   },
-};
+}
 
 const FermenterIndicator = ({ classes, isOn }: Props) => (
   <Avatar className={isOn ? classes.fermenterIconOn : classes.fermenterIconOff}>
     <FermenterIcon />
   </Avatar>
-);
+)
 
-export default compose(withStyles(fermenterIconStyles))(FermenterIndicator);
+export default compose(withStyles(fermenterIconStyles))(FermenterIndicator)
