@@ -1,23 +1,24 @@
 // @flow
 
-import type { Env } from '../../types/fermenter';
+import type { Env } from '../../types/fermenter'
 
-import * as React from 'react';
-import { connect } from 'react-redux';
+import * as React from 'react'
+import { connect } from 'react-redux'
 
-import { withStyles } from 'material-ui/styles';
-import Avatar from 'material-ui/Avatar';
-import Card, { CardContent } from 'material-ui/Card';
-import HumIcon from 'material-ui-icons/Cloud';
-import TempIcon from 'material-ui-icons/AcUnit';
-import Typography from 'material-ui/Typography';
+import { withStyles } from '@material-ui/core/styles'
+import Avatar from '@material-ui/core/Avatar'
+import Card from '@material-ui/core/Card'
+import CardContent from '@material-ui/core/CardContent'
+import Typography from '@material-ui/core/Typography'
+import HumIcon from '@material-ui/icons/Cloud'
+import TempIcon from '@material-ui/icons/AcUnit'
 
-import { compose } from 'ramda';
+import { compose } from 'ramda'
 
 type Props = {
   env: Env,
   classes: Object,
-};
+}
 
 const fermenterInfoStyles = {
   infoCard: {
@@ -32,7 +33,7 @@ const fermenterInfoStyles = {
     marginRight: 20,
     padding: 10,
   },
-};
+}
 
 const FermenterInfo = ({ env, classes }: Props) => (
   <Card className={classes.infoCard}>
@@ -55,9 +56,9 @@ const FermenterInfo = ({ env, classes }: Props) => (
       </CardContent>
     </div>
   </Card>
-);
+)
 
 export default compose(
   withStyles(fermenterInfoStyles),
   connect(state => ({ env: state.fermenter.env }))
-)(FermenterInfo);
+)(FermenterInfo)
