@@ -1,9 +1,9 @@
 // @flow
 
-import * as React from 'react';
-import { withStyles } from 'material-ui/styles';
-import Radio from 'material-ui/Radio';
-import Typography from 'material-ui/Typography';
+import * as React from 'react'
+import { withStyles } from '@material-ui/core/styles'
+import Radio from '@material-ui/core/Radio'
+import Typography from '@material-ui/core/Typography'
 
 const tempRangeControl = theme => ({
   group: {
@@ -13,24 +13,20 @@ const tempRangeControl = theme => ({
   selectors: {
     margin: `${theme.spacing.unit}px 0`,
   },
-});
+})
 
 type Props = {
   tempLimits: string,
   changeAction: Function,
   classes: Object,
-};
+}
 
-const FermenterTempRangeControl = ({
-  tempLimits,
-  changeAction,
-  classes,
-}: Props) => {
-  const tempLimitsAsStr = tempLimits ? tempLimits.toString() : '';
+const FermenterTempRangeControl = ({ tempLimits, changeAction, classes }: Props) => {
+  const tempLimitsAsStr = tempLimits ? tempLimits.toString() : ''
 
   const setTempRange = e =>
     /* Call action with new selected temp-range but convert numbers it from strings to floats first */
-    changeAction(e.target.value.split(',').map(n => parseFloat(n)));
+    changeAction(e.target.value.split(',').map(n => parseFloat(n)))
 
   return (
     <div className={classes.group}>
@@ -68,7 +64,7 @@ const FermenterTempRangeControl = ({
         <Typography>30-32</Typography>
       </div>
     </div>
-  );
-};
+  )
+}
 
-export default withStyles(tempRangeControl)(FermenterTempRangeControl);
+export default withStyles(tempRangeControl)(FermenterTempRangeControl)
