@@ -2,8 +2,8 @@
 
 /* Presentational component for a room-grouped list of switches */
 
-import type { KnxAddress, Prefs, Rooms } from '../../types';
-import * as React from 'react';
+import type { KnxAddress, Prefs, Rooms } from '../../types'
+import * as React from 'react'
 
 import {
   any,
@@ -21,15 +21,15 @@ import {
   propEq,
   sort,
   values,
-} from 'ramda';
+} from 'ramda'
 
-import { withStyles } from 'material-ui/styles';
-import Paper from 'material-ui/Paper';
-import List from 'material-ui/List';
-import Typography from 'material-ui/Typography';
+import { withStyles } from '@material-ui/core/styles'
+import Paper from '@material-ui/core/Paper'
+import List from '@material-ui/core/List'
+import Typography from '@material-ui/core/Typography'
 
-import AddressListItem from '../address/AddrListItem';
-import AddressSwitch from '../address/AddrItemSwitch';
+import AddressListItem from '../address/AddrListItem'
+import AddressSwitch from '../address/AddrItemSwitch'
 
 type Props = {
   addresses: Array<KnxAddress>,
@@ -37,7 +37,7 @@ type Props = {
   rooms: Rooms,
   onAddrSwitch: Function,
   classes: Object,
-};
+}
 
 const addrListStyles = theme => ({
   room: {
@@ -55,18 +55,12 @@ const addrListStyles = theme => ({
     background: theme.palette.background.paper,
     textAlign: 'auto',
   },
-});
+})
 
-const RoomList = ({
-  addresses,
-  prefs,
-  rooms,
-  onAddrSwitch,
-  classes,
-}: Props) => {
-  const hasRoom = room => any(equals(room), prefs.rooms);
-  const addressesWithRoom = keys(filter(hasRoom, pluck('room', addresses)));
-  const getRoomName = room => rooms[room].name;
+const RoomList = ({ addresses, prefs, rooms, onAddrSwitch, classes }: Props) => {
+  const hasRoom = room => any(equals(room), prefs.rooms)
+  const addressesWithRoom = keys(filter(hasRoom, pluck('room', addresses)))
+  const getRoomName = room => rooms[room].name
 
   const createRoomPanels = (addrLst, room) => (
     <Paper key={room} className={classes.room}>
@@ -78,16 +72,14 @@ const RoomList = ({
           <AddressListItem
             key={address.id}
             address={address}
-            addrSwitch={
-              <AddressSwitch switchAction={onAddrSwitch} address={address} />
-            }
+            addrSwitch={<AddressSwitch switchAction={onAddrSwitch} address={address} />}
           />
         ))}
       </List>
     </Paper>
-  );
+  )
 
-  const onlyActiveRooms = rooms => filter(any(propEq('value', 1)), rooms);
+  const onlyActiveRooms = rooms => filter(any(propEq('value', 1)), rooms)
 
   const addrLstByRoom = pipe(
     pick(addressesWithRoom),
@@ -97,9 +89,9 @@ const RoomList = ({
     prefs.showOnlyActive ? onlyActiveRooms : identity,
     mapObjIndexed(createRoomPanels),
     values /* NOTE: Make last result an array, otherwise React complains about an Object returned by #mapObjIndexed */
-  );
+  )
 
-  return <div>{addrLstByRoom(addresses)}</div>;
-};
+  return <div>{addrLstByRoom(addresses)}</div>
+}
 
-export default compose(withStyles(addrListStyles))(RoomList);
+export default compose(withStyles(addrListStyles))(RoomList)
