@@ -1,10 +1,10 @@
 // @flow
-import type { Action, AppState } from '../../types';
+import type { Action, AppState } from '../../types'
 
-import logger from 'debug';
-import { assoc, lensPath, set } from 'ramda';
+import logger from 'debug'
+import { assoc, lensPath, set } from 'ramda'
 
-const debug = logger('smtApp:appReducer');
+const debug = logger('smtApp:appReducer')
 
 const initialState: AppState = {
   rooms: {
@@ -57,45 +57,45 @@ const initialState: AppState = {
     state: 'undefined',
     error: 'all clear',
   },
-};
+}
 
 const reducer = (state: AppState = initialState, action: Action): AppState => {
   switch (action.type) {
     case 'SUBSCRIBE_TO_BUS_SUCCESS': {
-      return assoc('isSubscribedToBus', true, state);
+      return assoc('isSubscribedToBus', true, state)
     }
 
     case 'TOGGLE_SHOW_ONLY_ACTIVE': {
-      const valueLens = lensPath(['prefs', 'showOnlyActive']);
-      const newValue = action.toggleValue;
+      const valueLens = lensPath(['prefs', 'showOnlyActive'])
+      const newValue = action.toggleValue
 
-      return set(valueLens, newValue, state);
+      return set(valueLens, newValue, state)
     }
 
     case 'CHANGE_SELECTED_LIST_TAB': {
-      const valueLens = lensPath(['ui', 'selectedListTab']);
-      const newValue = action.value;
+      const valueLens = lensPath(['ui', 'selectedListTab'])
+      const newValue = action.value
 
-      return set(valueLens, newValue, state);
+      return set(valueLens, newValue, state)
     }
 
     case 'CONNECTION_STATE_UPDATE': {
-      const valueLens = lensPath(['connection', 'state']);
-      const { newState } = action;
+      const valueLens = lensPath(['connection', 'state'])
+      const { newState } = action
 
-      return set(valueLens, newState, state);
+      return set(valueLens, newState, state)
     }
 
     case 'CONNECTION_ERROR_UPDATE': {
-      const valueLens = lensPath(['connection', 'error']);
-      const { newError } = action;
+      const valueLens = lensPath(['connection', 'error'])
+      const { newError } = action
 
-      return set(valueLens, newError, state);
+      return set(valueLens, newError, state)
     }
 
     default:
-      return state;
+      return state
   }
-};
+}
 
-export default reducer;
+export default reducer
