@@ -1,24 +1,24 @@
 // @flow
-import type { Middleware, Reducers } from '../types';
-import createMiddleware from './create-middleware';
-import createReducer from './create-reducers';
-import { createStore } from 'redux';
+import type { Middleware, Reducers } from '../types'
+import createMiddleware from './create-middleware'
+import createReducer from './create-reducers'
+import { createStore } from 'redux'
 
 // Exact object type prevents typos.
 // https://flow.org/en/docs/types/objects/#toc-exact-object-types
 type Options = {|
   platformReducers?: Reducers,
   platformMiddlewares?: Middleware,
-|};
+|}
 
 const makeReducersHotReloadable = store => {
-  if (!module.hot || typeof module.hot.accept !== 'function') return;
+  if (!module.hot || typeof module.hot.accept !== 'function') return
   module.hot.accept('./create-reducers', () => {
     //     eslint-disable-next-line global-require
-    const createNextReducer = require('./create-reducers').default;
-    store.replaceReducer(createNextReducer());
-  });
-};
+    const createNextReducer = require('./create-reducers').default
+    store.replaceReducer(createNextReducer())
+  })
+}
 
 const createReduxStore = (
   // Global initial state; leaving keys out takes the initial state from their corresponding reducers instead (like
@@ -26,12 +26,12 @@ const createReduxStore = (
   initialState: Object = {},
   options?: Options
 ) => {
-  const reducers = createReducer();
-  const middleware = createMiddleware();
-  const store = createStore(reducers, initialState, middleware);
+  const reducers = createReducer()
+  const middleware = createMiddleware()
+  const store = createStore(reducers, initialState, middleware)
 
-  makeReducersHotReloadable(store);
-  return store;
-};
+  makeReducersHotReloadable(store)
+  return store
+}
 
-export default createReduxStore;
+export default createReduxStore
