@@ -1,6 +1,6 @@
 /* @flow */
 
-import type { Action, FermenterState } from '../../types/fermenter';
+import type { Action, FermenterState } from '../../types/fermenter'
 
 const initialState: FermenterState = {
   rts: {
@@ -37,20 +37,17 @@ const initialState: FermenterState = {
     switchOps: [],
     emergencies: [],
   },
-};
+}
 
-const reducer = (
-  state: FermenterState = initialState,
-  action: Action
-): FermenterState => {
+const reducer = (state: FermenterState = initialState, action: Action): FermenterState => {
   switch (action.type) {
     case 'PROCESS_STATE': {
-      return action.newState;
+      return action.newState
     }
 
     default:
-      return state;
+      return state
   }
-};
+}
 
-export default reducer;
+export default reducer
