@@ -1,8 +1,8 @@
 // @flow
-import type { CronJob, Task } from '../../types';
+import type { CronJob, Task } from '../../types'
 
-import shortid from 'shortid';
-import { assoc, lens, map, pipe, prop, replace, set, view } from 'ramda';
+import shortid from 'shortid'
+import { assoc, lens, map, pipe, prop, replace, set, view } from 'ramda'
 
 const jobTemplate: CronJob = {
   jobId: 'cronjobs/',
@@ -13,36 +13,36 @@ const jobTemplate: CronJob = {
   running: false,
   lastRun: null,
   tasks: null,
-};
+}
 
 function buildCronjobWithTasks(name: string, tasks: Array<Task>) {
-  const jobIdLens = lens(prop('jobId'), assoc('jobId'));
-  const orgJobId = view(jobIdLens)(jobTemplate);
+  const jobIdLens = lens(prop('jobId'), assoc('jobId'))
+  const orgJobId = view(jobIdLens)(jobTemplate)
   // Remove slash in scene-name and append unique id
-  const newJobId = set(
-    jobIdLens,
-    `${orgJobId}${replace(/\//, '', name)}-${shortid.generate()}`
-  );
+  const newJobId = set(jobIdLens, `${orgJobId}${replace(/\//, '', name)}-${shortid.generate()}`)
 
-  return pipe(newJobId, assoc('tasks', tasks))(jobTemplate);
+  return pipe(
+    newJobId,
+    assoc('tasks', tasks)
+  )(jobTemplate)
 }
 
 function scheduleJobForNow(job: CronJob) {
-  return assoc('at', 'now', job);
+  return assoc('at', 'now', job)
 }
 
 function invertTaskActions(tasks: Array<Task>) {
   const invert = act => {
     switch (act) {
       case 'on':
-        return 'off';
+        return 'off'
       case 'off':
-        return 'on';
+        return 'on'
       default:
-        return act;
+        return act
     }
-  };
-  return map(t => assoc('act', invert(prop('act', t)), t), tasks);
+  }
+  return map(t => assoc('act', invert(prop('act', t)), t), tasks)
 }
 
-export { buildCronjobWithTasks, scheduleJobForNow, invertTaskActions };
+export { buildCronjobWithTasks, scheduleJobForNow, invertTaskActions }
