@@ -1,7 +1,7 @@
-import { SheetsRegistry } from 'jss';
-import { createMuiTheme, createGenerateClassName } from 'material-ui/styles';
-import purple from 'material-ui/colors/purple';
-import green from 'material-ui/colors/green';
+import { SheetsRegistry } from 'jss'
+import { createMuiTheme, createGenerateClassName } from '@material-ui/core/styles'
+import purple from '@material-ui/core/colors/purple'
+import green from '@material-ui/core//colors/green'
 
 // A theme with custom primary and secondary color.
 // It's optional.
@@ -18,7 +18,7 @@ const theme = createMuiTheme({
       dark: green[700],
     },
   },
-});
+})
 
 function createPageContext() {
   return {
@@ -29,20 +29,20 @@ function createPageContext() {
     sheetsRegistry: new SheetsRegistry(),
     // The standard class name generator.
     generateClassName: createGenerateClassName(),
-  };
+  }
 }
 
 export default function getPageContext() {
   // Make sure to create a new context for every server-side request so that data
   // isn't shared between connections (which would be bad).
   if (!process.browser) {
-    return createPageContext();
+    return createPageContext()
   }
 
   // Reuse context on the client-side.
   if (!global.__INIT_MATERIAL_UI__) {
-    global.__INIT_MATERIAL_UI__ = createPageContext();
+    global.__INIT_MATERIAL_UI__ = createPageContext()
   }
 
-  return global.__INIT_MATERIAL_UI__;
+  return global.__INIT_MATERIAL_UI__
 }
