@@ -1,45 +1,34 @@
+// @flow
+
 // Right now this code is mostly about making Material-UI work with nextjs - see
 // https://github.com/mui-org/material-ui/blob/v1-beta/examples/nextjs/pages/_document.js
 
-import React from 'react';
-import Document, { Head, Main, NextScript } from 'next/document';
-import JssProvider from 'react-jss/lib/JssProvider';
-import flush from 'styled-jsx/server';
-import getPageContext from '../lib/shared/getPageContext';
+import React from 'react'
+import PropTypes from 'prop-types'
+import Document, { Head, Main, NextScript } from 'next/document'
+import flush from 'styled-jsx/server'
 
 class MyDocument extends Document {
   render() {
-    const { pageContext } = this.props;
+    const { pageContext } = this.props
 
     return (
       <html lang="en" dir="ltr">
         <Head>
-          <title>CjK&#39;s smart-home</title>
+          <title>CjK's smarthome</title>
           <meta charSet="utf-8" />
           {/* Use minimum-scale=1 to enable GPU rasterization */}
-          <meta
-            name="viewport"
-            content={
-              'user-scalable=0, initial-scale=1, ' +
-              'minimum-scale=1, width=device-width, height=device-height'
-            }
-          />
+          <meta name="viewport" content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no" />
           {/* PWA primary color */}
-          <meta
-            name="theme-color"
-            content={pageContext.theme.palette.primary.main}
-          />
-          <link
-            rel="stylesheet"
-            href="https://fonts.googleapis.com/css?family=Roboto:300,400,500"
-          />
+          <meta name="theme-color" content={pageContext.theme.palette.primary.main} />
+          <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500" />
         </Head>
         <body>
           <Main />
           <NextScript />
         </body>
       </html>
-    );
+    )
   }
 }
 
@@ -47,34 +36,44 @@ MyDocument.getInitialProps = ctx => {
   // Resolution order
   //
   // On the server:
-  // 1. page.getInitialProps
-  // 2. document.getInitialProps
-  // 3. page.render
-  // 4. document.render
+  // 1. app.getInitialProps
+  // 2. page.getInitialProps
+  // 3. document.getInitialProps
+  // 4. app.render
+  // 5. page.render
+  // 6. document.render
   //
   // On the server with error:
-  // 2. document.getInitialProps
+  // 1. document.getInitialProps
+  // 2. app.render
   // 3. page.render
   // 4. document.render
   //
   // On the client
-  // 1. page.getInitialProps
-  // 3. page.render
+  // 1. app.getInitialProps
+  // 2. page.getInitialProps
+  // 3. app.render
+  // 4. page.render
 
-  // Get the context of the page to collected side effects.
-  const pageContext = getPageContext();
-  const page = ctx.renderPage(Component => props => (
-    <JssProvider
-      registry={pageContext.sheetsRegistry}
-      generateClassName={pageContext.generateClassName}
-    >
-      <Component pageContext={pageContext} {...props} />
-    </JssProvider>
-  ));
+  // Render app and page and get the context of the page with collected side effects.
+  let pageContext
+  const page = ctx.renderPage(Component => {
+    const WrappedComponent = props => {
+      pageContext = props.pageContext
+      return <Component {...props} />
+    }
+
+    WrappedComponent.propTypes = {
+      pageContext: PropTypes.object.isRequired,
+    }
+
+    return WrappedComponent
+  })
 
   return {
     ...page,
     pageContext,
+    // Styles fragment is rendered after the app and page rendering finish.
     styles: (
       <React.Fragment>
         <style
@@ -87,7 +86,7 @@ MyDocument.getInitialProps = ctx => {
         {flush() || null}
       </React.Fragment>
     ),
-  };
-};
+  }
+}
 
-export default MyDocument;
+export default MyDocument
