@@ -1,19 +1,18 @@
 // @flow
-import type { FermenterState } from '../types/fermenter';
-import type { Dispatch } from '../types';
 
-import * as React from 'react';
-import * as fermenterActions from '../lib/fermenter/actions';
-import createStore from '../lib/create-store';
-import withRedux from 'next-redux-wrapper';
-import withRoot from '../components/hocs/withRoot';
+import type { FermenterState } from '../types/fermenter'
+import type { Dispatch } from '../types'
 
-import AppBar from '../components/AppBar';
-import FermenterInfo from '../components/fermenter/FermenterInfo';
-import FermenterControl from '../components/fermenter/FermenterControl';
-import Paper from 'material-ui/Paper';
+import * as React from 'react'
+import * as fermenterActions from '../lib/fermenter/actions'
+import { connect } from 'react-redux'
 
-import { compose } from 'ramda';
+import AppBar from '../components/AppBar'
+import FermenterInfo from '../components/fermenter/FermenterInfo'
+import FermenterControl from '../components/fermenter/FermenterControl'
+import Paper from '@material-ui/core/Paper'
+
+import { compose } from 'ramda'
 
 type Props = {
   fermenter: FermenterState,
@@ -23,49 +22,48 @@ type Props = {
   sendFermenterTempLimits: Function,
   subscribeToState: Function,
   unsubscribeToState: Function,
-};
+}
 
 const styles = {
   fermenterRoot: {
     margin: 20,
     padding: 10,
   },
-};
+}
 
 class FermenterPage extends React.Component<void, Props> {
   // TODO: No server-side-rendering supported yet :(
   static async getInitialProps(ctx) {
-    const { store } = ctx;
-    return store;
+    const { store } = ctx
+    return store
   }
 
   componentDidMount() {
-    this.props.subscribeToState();
+    this.props.subscribeToState()
   }
 
   componentWillUnmount() {
-    this.props.unsubscribeToState();
+    this.props.unsubscribeToState()
   }
 
   render() {
-    const { sendFermenterCmd, sendFermenterTempLimits } = this.props;
+    const { sendFermenterCmd, sendFermenterTempLimits } = this.props
 
     return (
       <div className="app">
         <AppBar />
         <Paper style={styles.fermenterRoot}>
-          <FermenterControl
-            sendFermenterCmd={sendFermenterCmd}
-            sendFermenterTempLimits={sendFermenterTempLimits}
-          />
+          <FermenterControl sendFermenterCmd={sendFermenterCmd} sendFermenterTempLimits={sendFermenterTempLimits} />
           <FermenterInfo />
         </Paper>
       </div>
-    );
+    )
   }
 }
 
 export default compose(
-  withRedux(createStore, null, fermenterActions),
-  withRoot
-)(FermenterPage);
+  connect(
+    null,
+    fermenterActions
+  )
+)(FermenterPage)
