@@ -1,12 +1,7 @@
-import * as React from 'react';
+import * as React from 'react'
 
-import createStore from '../lib/create-store';
-import withRedux from 'next-redux-wrapper';
-import withRoot from '../components/hocs/withRoot';
-import withBusSubscribe from '../components/hocs/withBusSubscribe';
-import AppBar from '../components/AppBar';
-import Dashboard from '../components/dashboard';
-import { compose } from 'ramda';
+import AppBar from '../components/AppBar'
+import Dashboard from '../components/dashboard'
 
 const IndexPage = () => (
   <div className="app">
@@ -25,8 +20,6 @@ const IndexPage = () => (
       }
     `}</style>
   </div>
-);
+)
 
-export default compose(withRedux(createStore), withBusSubscribe, withRoot)(
-  IndexPage
-);
+export default IndexPage
