@@ -8,42 +8,42 @@ export type Env = {
   isValid: boolean,
   errors: number,
   iterations: number,
-};
+}
 
 type Device = {
   isOn: boolean,
   shouldSwitchTo: ?string,
   willSwitch: boolean,
-};
+}
 
 export type Devices = {
   heater: Device,
   humidifier: Device,
-};
+}
 
 export type SwitchOp = {
   device: ?string,
   to: ?string,
   at: ?number,
-};
+}
 
 export type Emergency = {
   device: ?string,
   sensor: ?string,
   at: ?number,
-};
+}
 
 type Notification = {
   level: string,
   msg: ?string,
-};
+}
 
 export type History = {
   switchOps: Array<SwitchOp>,
   emergencies: Array<Emergency>,
-};
+}
 
-export type EnvLimits = Array<number>;
+export type EnvLimits = Array<number>
 
 export type RunTimeState = {
   active: ?boolean,
@@ -54,14 +54,14 @@ export type RunTimeState = {
   tempLimits: ?EnvLimits,
   humidityLimits: ?EnvLimits,
   notifications: ?Array<Notification>,
-};
+}
 
 export type FermenterState = {
   rts: RunTimeState,
   env: Env,
   devices: Devices,
   history: History,
-};
+}
 
 export type Action =
   | { type: 'SUBSCRIBE_TO_STATE' }
@@ -71,4 +71,4 @@ export type Action =
   | {
       type: 'SEND_FERMENTER_TEMPLIMITS',
       tempLimits: EnvLimits,
-    };
+    }
